move RailsBestPractices to RailsBestPractices::Analyzer
